+++ lib/codesake/dawn/core.rb
@@ -1,3 +1,5 @@
+require "yaml"
+
 module Codesake
   module Dawn
     class Core
@@ -13,15 +15,16 @@ module Codesake
         puts "\t$ dawn -C --json a_sinatra_webapp_directory"
         puts "\t$ dawn --ascii-tabular-report my_rails_blog_ecommerce"
         puts "\t$ dawn --html -F my_report.html my_rails_blog_ecommerce"
-        printf "\n   -r, --rails\t\t\t\t\tforce dawn to consider the target a rails application" 
-        printf "\n   -s, --sinatra\t\t\t\tforce dawn to consider the target a sinatra application" 
-        printf "\n   -p, --padrino\t\t\t\tforce dawn to consider the target a padrino application" 
+        printf "\n   -r, --rails\t\t\t\t\tforce dawn to consider the target a rails application"
+        printf "\n   -s, --sinatra\t\t\t\tforce dawn to consider the target a sinatra application"
+        printf "\n   -p, --padrino\t\t\t\tforce dawn to consider the target a padrino application"
         printf "\n   -G, --gem-lock\t\t\t\tforce dawn to scan only for vulnerabilities affecting dependencies in Gemfile.lock"
         printf "\n   -a, --ascii-tabular-report\t\t\tcause dawn to format findings using table in ascii art"
         printf "\n   -j, --json\t\t\t\t\tcause dawn to format findings using json"
         printf "\n   -C, --count-only\t\t\t\tdawn will only count vulnerabilities (useful for scripts)"
         printf "\n   -z, --exit-on-warn\t\t\t\tdawn will return number of found vulnerabilities as exit code"
-        printf "\n   -F, --file\t\t\t\t\ttells dawn to write output to filename"
+        printf "\n   -F, --file filename\t\t\t\ttells dawn to write output to filename"
+        printf "\n   -c, --config-file filename\t\t\ttells dawn to load configuration from filename"
         printf "\n\nDisable security check family\n"
         printf "\n       --disable-cve-bulletins\t\t\tdisable all CVE security checks"
         printf "\n       --disable-code-quality\t\t\tdisable all code quality checks"
@@ -38,7 +41,7 @@ module Codesake
         printf "\n   -v, --version\t\t\t\tshow version information"
         printf "\n   -h, --help\t\t\t\t\tshow this help\n"
 
-        true 
+        true
       end
 
       def self.dump_knowledge_base(verbose = false)
@@ -70,7 +73,7 @@ module Codesake
 
         a = []
         my_dir = Dir.pwd
-        Dir.chdir(File.dirname(gemfile_lock)) 
+        Dir.chdir(File.dirname(gemfile_lock))
         raise ArgumentError.new("can't read #{gemfile_lock}") unless File.readable?(File.basename(gemfile_lock))
 
         lockfile = Bundler::LockfileParser.new(Bundler.read_file(File.basename(gemfile_lock)))
@@ -90,7 +93,7 @@ module Codesake
         raise ArgumentError.new("you must set target directory") if target.nil?
 
         my_dir = Dir.pwd
-        Dir.chdir(target) 
+        Dir.chdir(target)
         raise ArgumentError.new("no Gemfile.lock in #{target}") unless File.exist?("Gemfile.lock")
 
         lockfile = Bundler::LockfileParser.new(Bundler.read_file("Gemfile.lock"))
@@ -100,11 +103,59 @@ module Codesake
           return Codesake::Dawn::Padrino.new(target)  if s.name == "padrino"
         end
 
-        return Codesake::Dawn::Sinatra.new(target) 
+        return Codesake::Dawn::Sinatra.new(target)
       end
 
       def self.is_good_target?(target)
         (File.exist?(target) and File.directory?(target))
+      end
+
+      def self.find_conf(create_if_none = false)
+        conf_name  = 'codesake-dawn.yaml'
+        path_order = [
+          './',
+          '~/',
+          '/usr/local/etc/',
+        ]
+        path_order.each do |p|
+          fn = p + conf_name if p.start_with?('/')
+          # if outside $HOME the config file must be hidden
+          fn = File.expand_path(p) + '/.'+conf_name if ! p.start_with?('/')
+          return fn if File.exist?(fn)
+        end
+
+        # Codesake::Dawn didn't find a config file.
+        # If create_if_none flag is set to false, than I'll return nil so the
+        # read_conf method will return the default configuration
+        return nil unless create_if_none
+
+        # If create_if_none flag is set to true, than I'll create a config file
+        # on the current directory with the default configuration.
+        conf = {"config"=>{:verbose=>false, :output=>"console", :mvc=>"", :gemfile_scan=>false, :gemfile_name=>"", :filename=>nil, :debug=>false, :exit_on_warn => false, :enabled_checks=> Codesake::Dawn::Kb::BasicCheck::ALLOWED_FAMILIES}}
+
+        File.open(File.expand_path('~') +'/.'+conf_name, 'w') do |f|
+          f.write(YAML.dump(conf))
+        end
+      end
+
+      def self.read_conf(file=nil)
+
+        conf = {:verbose=>false, :output=>"console", :mvc=>"", :gemfile_scan=>false, :gemfile_name=>"", :filename=>nil, :debug=>false, :exit_on_warn => false, :enabled_checks=> Codesake::Dawn::Kb::BasicCheck::ALLOWED_FAMILIES}
+        return conf if file.nil?
+        return conf if ! File.exist?(file)
+
+        c = YAML.load_file(file)
+
+        cf = c["config"]
+        cc = cf["enabled_checks"]
+        # TODO
+        # I must add some sanity check here
+        conf[:verbose] = cf["verbose"] unless cf["verbose"].nil?
+        conf[:debug] = cf["debug"] unless cf["debug"].nil?
+        conf[:output] = cf["output"] unless cf["output"].nil?
+        conf[:enabled_checks] = cc unless cc.nil?
+
+        return conf
       end
     end
   end